Move getDirSize function from lumine-fileinfo to libLumina.
Now the function can be used everywhere. At this moment, static variables are controled
by a QHash with a uuid as index for differentiate who calls the function.

I check that this QHash is not needed because the static variable are local to every process
but I must be sure that this is what standard says.